@@ -1,6 +1,6 @@
 plugins {
-    id "pl.allegro.tech.build.axion-release" version "1.14.2"
-    id "org.springframework.boot" version "2.7.5"
+    id "pl.allegro.tech.build.axion-release" version "1.14.3"
+    id "org.springframework.boot" version "3.0.1"
     id "org.sonarqube" version "3.5.0.2730"
     id 'com.google.cloud.tools.jib' version '3.3.1'
 }
@@ -27,8 +27,8 @@ ext.bootstrapVersion = "5.2.2"
 ext.scmUrl = "git@github.com:gessnerfl/fake-smtp-server.git"
 ext.projectUrl = "https://github.com/gessnerfl/fake-smtp-server"
 
-sourceCompatibility = 11
-targetCompatibility = 11
+sourceCompatibility = 17
+targetCompatibility = 17
 
 repositories {
     mavenCentral()
@@ -39,13 +39,13 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-thymeleaf")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+    implementation("org.springframework.boot:spring-boot-starter-validation")
     implementation("org.springframework.boot:spring-boot-starter-mail")
     implementation("commons-io:commons-io:2.11.0")
-    implementation("org.subethamail:subethasmtp:3.1.7")
-    implementation("org.flywaydb:flyway-core:8.5.13")
+    implementation('org.flywaydb:flyway-core')
     implementation("org.jsoup:jsoup:1.15.3")
-    implementation("com.sun.xml.bind:jaxb-impl:4.0.1")
-    implementation("org.springdoc:springdoc-openapi-ui:1.6.12")
+    implementation("org.glassfish.jaxb:jaxb-runtime:4.0.1")
+    implementation('org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.0')
 
     implementation("org.webjars:bootstrap:$bootstrapVersion")
     implementation("org.webjars:material-design-icons:4.0.0")
@@ -59,7 +59,6 @@ dependencies {
 
 java {
     withSourcesJar()
-    withJavadocJar()
 }
 
 gradle.projectsEvaluated {
