@@ -3,7 +3,7 @@ package de.gessnerfl.fakesmtp.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.net.InetAddress;
 
 @Component
