@@ -13,8 +13,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.ServletContext;
-import javax.validation.constraints.Min;
+import jakarta.servlet.ServletContext;
+import jakarta.validation.constraints.Min;
 import java.util.Collections;
 import java.util.List;
 
