@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.subethamail.smtp.helper.SimpleMessageListener;
+import de.gessnerfl.fakesmtp.server.smtp.helper.SimpleMessageListener;
 
 import java.io.IOException;
 import java.io.InputStream;
